+++ components/auth/Confirmation/Confirmation.test.js
@@ -6,9 +6,7 @@ import userEvent from "@testing-library/user-event";
 
 jest.mock("../../../lib/hooks/auth");
 
-// TODO: skipping test for now. This component will probably be removed in the near future.
-
-describe.skip("confirmation component", () => {
+describe("confirmation component", () => {
   let confirmMock;
   let resendConfirmationMock;
   beforeEach(() => {
