+++ components/auth/Confirmation/Confirmation.tsx
@@ -1,18 +1,11 @@
 import React, { ReactElement, useState } from "react";
-import { useRouter } from "next/router";
-import { Formik, FormikHelpers } from "formik";
-import { TextInput, Label, Alert, Description } from "@components/forms";
+import { Formik } from "formik";
+import { Button, TextInput, Label, Alert, ErrorListItem } from "@components/forms";
+import { useConfirm } from "@lib/hooks/auth";
 import { useTranslation } from "next-i18next";
 import * as Yup from "yup";
 import Link from "next/link";
 import { ErrorStatus } from "@components/forms/Alert/Alert";
-import axios from "axios";
-import { getCsrfToken } from "next-auth/react";
-import { hasError } from "@lib/hasError";
-import { fetchWithCsrfToken } from "@lib/hooks/auth/fetchWithCsrfToken";
-import { useAuthErrors } from "@lib/hooks/auth/useAuthErrors";
-import { logMessage } from "@lib/logger";
-import { Button } from "@components/globals";
 
 interface ConfirmationProps {
   username: string;
@@ -27,121 +20,20 @@ export const Confirmation = ({
   confirmationCallback,
   shouldSignIn = true,
 }: ConfirmationProps): ReactElement => {
-  const router = useRouter();
-  const { t } = useTranslation(["signup", "cognito-errors"]);
+  const { confirm, resendConfirmationCode, authErrorsState, authErrorsReset } = useConfirm();
   const [showSentReconfirmationToast, setShowSentReconfirmationToast] = useState(false);
-  const [authErrorsState, { authErrorsReset, handleErrorById }] = useAuthErrors();
-
-  const confirm = async (
-    {
-      confirmationCode,
-      confirmationCallback,
-      shouldSignIn = true,
-    }: {
-      confirmationCode: string;
-      confirmationCallback: () => void;
-      shouldSignIn: boolean;
-    },
-    { setSubmitting, setErrors }: FormikHelpers<{ confirmationCode: string }>
-  ) => {
-    authErrorsReset();
-
-    let confirmationSuccess = true;
-    try {
-      await fetchWithCsrfToken("/api/signup/confirm", {
-        username: username,
-        confirmationCode,
-      });
-    } catch (err) {
-      logMessage.error(err);
-      confirmationSuccess = false;
-
-      if (hasError("CodeMismatchException", err)) {
-        setErrors({ confirmationCode: t("CodeMismatchException") });
-        return;
-      }
-      if (hasError("ExpiredCodeException", err)) {
-        setErrors({ confirmationCode: t("ExpiredCodeException") });
-        return;
-      }
-      handleErrorById("InternalServiceException");
-    } finally {
-      // set the formik submitting state to false
-      setSubmitting(false);
-    }
-
-    // end the execution of the function if the confirmation did not succeed
-    if (!confirmationSuccess) return;
-
-    // if automated sign up is disabled. call the confirmation callback and end the execution
-    if (!shouldSignIn) {
-      confirmationCallback();
-      return;
-    }
-
-    // try and sign the user in automatically if shouldSignIn is true, otherwise just
-    // call the passed in confirmationCallback
-    try {
-      const { data } = await axios({
-        url: "/api/auth/signin/cognito",
-        method: "POST",
-        headers: {
-          "Content-Type": "application/x-www-form-urlencoded",
-        },
-        data: new URLSearchParams({
-          username: username,
-          password: password,
-          csrfToken: (await getCsrfToken()) ?? "noToken",
-        }),
-        timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
-      });
-
-      if (data?.error) {
-        const error = data.error;
-        logMessage.error(error);
-        if (hasError(["UserNotFoundException", "NotAuthorizedException"], error)) {
-          handleErrorById("UsernameOrPasswordIncorrect");
-        } else {
-          handleErrorById("InternalServiceException");
-        }
-      } else if (data?.ok) {
-        await router.push("/auth/policy?referer=/signup/account-created");
-      }
-    } catch (err) {
-      logMessage.error(err);
-      handleErrorById("InternalServiceException");
-      // Internal error on sign in, not confirmation, so redirect to login page
-      await router.push("/auth/login");
-    } finally {
-      confirmationCallback();
-    }
-  };
-
-  const resendConfirmationCode = async (username: string) => {
-    authErrorsReset();
-    try {
-      await fetchWithCsrfToken("/api/signup/resendconfirmation", { username });
-    } catch (err) {
-      logMessage.error(err);
-      if (hasError("TooManyRequestsException", err)) {
-        handleErrorById("TooManyRequestsException");
-        return;
-      }
-      handleErrorById("InternalServiceException");
-      return err;
-    }
-  };
+  const { t } = useTranslation(["signup", "cognito-errors", "common"]);
 
   const validationSchema = Yup.object().shape({
-    confirmationCode: Yup.string()
-      .required(t("signUpConfirmation.fields.confirmationCode.error.notEmpty"))
-      .matches(/^[0-9a-z]{5}?$/i, t("signUpConfirmation.fields.confirmationCode.error.length")),
+    confirmationCode: Yup.number()
+      .typeError(t("signUpConfirmation.fields.confirmationCode.error.number"))
+      .required(t("input-validation.required", { ns: "common" })),
   });
 
   return (
     <Formik
       validationSchema={validationSchema}
-      initialValues={{ confirmationCode: "" }}
+      initialValues={{ username: username, password: password, confirmationCode: "" }}
       validateOnBlur={false}
       validateOnChange={false}
       onSubmit={async (values, formikHelpers) => {
@@ -156,7 +48,7 @@ export const Confirmation = ({
         );
       }}
     >
-      {({ handleSubmit }) => (
+      {({ handleSubmit, errors }) => (
         <>
           {showSentReconfirmationToast && !authErrorsState?.title && (
             <Alert
@@ -186,7 +78,7 @@ export const Confirmation = ({
               ) : undefined}
             </Alert>
           )}
-          {/* {Object.keys(errors).length > 0 && !authErrorsState.isError && (
+          {Object.keys(errors).length > 0 && !authErrorsState.isError && (
             <Alert
               type={ErrorStatus.ERROR}
               validation={true}
@@ -206,9 +98,11 @@ export const Confirmation = ({
                 })}
               </ol>
             </Alert>
-          )} */}
-          <h1 className="border-0">{t("signUpConfirmation.title")}</h1>
-          <p className="mb-10 -mt-6">{t("signUpConfirmation.emailHasBeenSent")}</p>
+          )}
+          <h1>{t("signUpConfirmation.title")}</h1>
+          <p className="mb-10 -mt-6">
+            {t("signUpConfirmation.emailHasBeenSent")}&nbsp;{username}
+          </p>
           <form id="confirmation" method="POST" onSubmit={handleSubmit} noValidate>
             <div className="focus-group">
               <Label
@@ -219,36 +113,29 @@ export const Confirmation = ({
               >
                 {t("signUpConfirmation.fields.confirmationCode.label")}
               </Label>
-              <Description className="text-p text-black-default" id={"confirmationCode-hint"}>
-                {t("signUpConfirmation.fields.confirmationCode.description")}
-              </Description>
               <TextInput
                 className="h-10 w-36 rounded"
                 type="text"
                 id="confirmationCode"
                 name="confirmationCode"
-                ariaDescribedBy="confirmationCode-hint"
                 required
               />
             </div>
-            <Button theme="primary" type="submit">
+            <Button className="gc-button--blue" type="submit">
               {t("signUpConfirmation.confirmButton")}
             </Button>
-            <div className="flex mt-10">
-              <button
-                type="button"
-                className="block shadow-none bg-transparent text-blue-dark hover:text-blue-hover underline border-0 mr-8"
-                onClick={async () => {
-                  const error = await resendConfirmationCode(username.current);
-                  if (!error) {
-                    setShowSentReconfirmationToast(true);
-                  }
-                }}
-              >
-                {t("signUpConfirmation.resendConfirmationCodeButton")}
-              </button>
-              <Link href={"/form-builder/support"}>{t("signUpConfirmation.help")}</Link>
-            </div>
+            <button
+              type="button"
+              className="block my-10 shadow-none bg-transparent text-blue-dark hover:text-blue-hover underline border-0"
+              onClick={async () => {
+                const error = await resendConfirmationCode(username);
+                if (!error) {
+                  setShowSentReconfirmationToast(true);
+                }
+              }}
+            >
+              {t("signUpConfirmation.resendConfirmationCodeButton")}
+            </button>
           </form>
         </>
       )}
